avoid size_t for printing

+++ src/platform/silabs/tracing/SilabsPowerTracing.cpp
@@ -127,9 +127,9 @@ CHIP_ERROR SilabsPowerTracing::OutputPowerManagerTraces()
 
     if (mEnergyTraces != nullptr && mEnergyTraceCount > 0)
     {
-        for (size_t i = 0; i < mEnergyTraceCount; i++)
+        for (uint32_t i = 0; i < mEnergyTraceCount; i++)
         {
-            ChipLogProgress(DeviceLayer, "%u | %lu | EM%d", i, mEnergyTraces[i].mEntryTime, mEnergyTraces[i].mEnergyMode);
+            ChipLogProgress(DeviceLayer, "%lu | %lu | EM%d", i, mEnergyTraces[i].mEntryTime, mEnergyTraces[i].mEnergyMode);
             // Delay so the output is not mangled or skipped.
             // 5 (ticks) is enough for UART, but only 1 is required for RTT.
             // No delay results in missed or mangled output for both.
